+++ examples/record.rs
@@ -1,34 +1,38 @@
 // This example records audio for 5 seconds and writes to a raw PCM file.
 
+// Setup async main
+include!(concat!(env!("OUT_DIR"), "/main.rs"));
+
 use fon::{mono::Mono32, Audio, Frame};
-use pasts::{exec, wait};
+use pasts::{prelude::*, Join};
 use wavy::{Microphone, MicrophoneStream};
 
-/// An event handled by the event loop.
-enum Event<'a> {
-    /// Microphone has recorded some audio.
-    Record(MicrophoneStream<'a, Mono32>),
-}
-
 /// Shared state between tasks on the thread.
-struct State {
+struct App<'a> {
+    /// Handle to the mono microphone
+    microphone: &'a mut Microphone<1>,
     /// Temporary buffer for holding real-time audio samples.
     buffer: Audio<Mono32>,
 }
 
-impl State {
+impl App<'_> {
     /// Event loop.  Return false to stop program.
-    fn event(&mut self, event: Event<'_>) {
-        match event {
-            Event::Record(microphone) => {
-                //println!("Recording");
-                self.buffer.extend(microphone);
-                if self.buffer.len() >= 48_000 * 10 {
-                    write_pcm(&self.buffer);
-                    std::process::exit(0);
-                }
-            }
+    fn record(&mut self, stream: MicrophoneStream<Mono32>) -> Poll<()> {
+        self.buffer.extend(stream);
+        if self.buffer.len() >= 48_000 * 10 {
+            return Ready(());
         }
+        Pending
+    }
+
+    async fn main(_executor: Executor) {
+        let buffer = Audio::with_silence(48_000, 0);
+        let microphone = &mut Microphone::default();
+        let mut app = App { buffer, microphone };
+
+        Join::new(&mut app).on(|s| s.microphone, App::record).await;
+
+        write_pcm(&app.buffer);
     }
 }
 
@@ -40,16 +44,4 @@ fn write_pcm(buffer: &Audio<Mono32>) {
         pcm.extend(sample.to_le_bytes().iter());
     }
     std::fs::write("pcm.raw", pcm.as_slice()).expect("Failed to write file");
-}
-
-/// Program start.
-fn main() {
-    let mut state = State {
-        buffer: Audio::with_silence(48_000, 0),
-    };
-    let mut microphone = Microphone::default();
-
-    exec!(state.event(wait! {
-        Event::Record(microphone.record().await),
-    }))
 }